apps/quest-controller/requirements.txt
fastapi
uvicorn[standard]
databricks-sdk
databricks-sql-connector
